Combine read and write demos
