+++ test/Makefile
@@ -5,10 +5,7 @@ CFLAGS=-g -std=gnu99 -I. -I$(SQLITE) -L$(SQLITE)/.libs -lsqlite3
 
 all: vfsbench testnfsvfs nfsv4_read_demo nfsv4_write_demo
 
-nfsv4_read_demo: nfsv4_read_demo.c
-	gcc -Wall -g -std=gnu99 -I../nfsv4 -L../nfsv4 -lnfs4 $< -o $@
-
-nfsv4_write_demo: nfsv4_write_demo.c
+nfsv4_read_write_test: nfsv4_read_write_test.c
 	gcc -Wall -g -std=gnu99 -I../nfsv4 -L../nfsv4 -lnfs4 $< -o $@
 
 vfsbench: vfsbench.c check-env
